refactored nylas service
