+++ backend/app/api/routes/clerk.py
@@ -6,7 +6,6 @@ from svix.webhooks import Webhook, WebhookVerificationError
 
 from app.services.clerk import post_user, get_clerk_private_metadata
 from app.models.users import UserMetadataResponse
-from app.api.routes.nylas_service import send_notification_email
 
 logger = logging.getLogger(__name__)
